Support BINARY/MULTISTATE/ALPHANUMERIC data types in aggregate() method

+++ Core/src/com/serotonin/m2m2/rt/script/ValueChangeCounterWrapper.java
@@ -60,7 +60,7 @@ public class ValueChangeCounterWrapper {
     	return statistics.getLastTime();
     }
 
-    public int getCount() {
+    public long getCount() {
         return statistics.getCount();
     }
